+++ sdk-incubator-lib/src/main/java/cash/z/ecc/android/sdk/model/ZecSend.kt
@@ -2,7 +2,12 @@ package cash.z.ecc.android.sdk.model
 
 import cash.z.ecc.android.sdk.Synchronizer
 
-data class ZecSend(val destination: WalletAddress, val amount: Zatoshi, val memo: Memo) {
+data class ZecSend(
+    val destination: WalletAddress,
+    val amount: Zatoshi,
+    val memo: Memo,
+    val proposal: Proposal?
+) {
     companion object
 }
 
@@ -19,11 +24,14 @@ suspend fun Synchronizer.send(
     spendingKey
 )
 
+/**
+ * This is just a syntactic sugar function for [Synchronizer.proposeTransfer]
+ */
 suspend fun Synchronizer.proposeSend(
-    spendingKey: UnifiedSpendingKey,
+    account: Account,
     send: ZecSend
 ) = proposeTransfer(
-    spendingKey.account,
+    account,
     send.destination.address,
     send.amount,
     send.memo.value